Remove need to create defensive copy by internalizing conversion of collection to class consuming it
