+++ platforms/core-runtime/build-operations-trace/src/main/java/org/gradle/internal/operations/trace/BuildOperationRecord.java
@@ -27,6 +27,7 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 import static com.google.common.collect.Lists.transform;
+import static org.gradle.internal.Cast.uncheckedCast;
 
 public final class BuildOperationRecord {
 
@@ -59,7 +60,7 @@ public final class BuildOperationRecord {
         @Nullable Map<String, ?> result,
         @Nullable String resultClassName,
         @Nullable String failure,
-        List<Progress> progress,
+        List<SerializedOperationProgress> progress,
         List<BuildOperationRecord> children
     ) {
         this.id = id;
@@ -74,10 +75,23 @@ public final class BuildOperationRecord {
         this.failure = failure;
         this.children = children;
 
-        this.progress = new ArrayList<>(progress);
+        this.progress = convertProgressEvents(progress);
     }
 
-    Map<String, ?> toSerializable() {
+    private static List<BuildOperationRecord.Progress> convertProgressEvents(List<SerializedOperationProgress> toConvert) {
+        List<BuildOperationRecord.Progress> progresses = new ArrayList<>();
+        for (SerializedOperationProgress progress : toConvert) {
+            Map<String, ?> progressDetailsMap = uncheckedCast(progress.details);
+            progresses.add(new BuildOperationRecord.Progress(
+                progress.time,
+                progressDetailsMap,
+                progress.detailsClassName
+            ));
+        }
+        return progresses;
+    }
+
+    /* package */ Map<String, ?> toSerializable() {
         Map<String, Object> map = new LinkedHashMap<>();
         map.put("displayName", displayName);
