+++ platforms/core-runtime/build-operations-trace/src/main/java/org/gradle/internal/operations/trace/BuildOperationTrace.java
@@ -395,7 +395,7 @@ public class BuildOperationTrace implements Stoppable {
                             resultMap == null ? null : Collections.unmodifiableMap(resultMap),
                             finish.resultClassName,
                             finish.failureMsg,
-                            convertProgressEvents(pending.progress),
+                            pending.progress,
                             BuildOperationRecord.ORDERING.immutableSortedCopy(children)
                         );
 
@@ -428,7 +428,7 @@ public class BuildOperationTrace implements Stoppable {
                     -1L, null,
                     "Dangling pending operations",
                     0L, 0L, null, null, null, null, null,
-                    convertProgressEvents(danglingProgress),
+                    danglingProgress,
                     Collections.emptyList()
                 ));
             }
@@ -438,19 +438,6 @@ public class BuildOperationTrace implements Stoppable {
             throw UncheckedException.throwAsUncheckedException(e);
         }
 
-    }
-
-    private static List<BuildOperationRecord.Progress> convertProgressEvents(List<SerializedOperationProgress> toConvert) {
-        List<BuildOperationRecord.Progress> progresses = new ArrayList<>();
-        for (SerializedOperationProgress progress : toConvert) {
-            Map<String, ?> progressDetailsMap = uncheckedCast(progress.details);
-            progresses.add(new BuildOperationRecord.Progress(
-                progress.time,
-                progressDetailsMap,
-                progress.detailsClassName
-            ));
-        }
-        return progresses;
     }
 
     private static File logFile(String basePath) {
